Fix several bugs in jump table code generation.
Now the stack push will correctly resume execution at next instruction
at the function after carrying out the push. Same for the stack pop. We
return to the function after the stack pop and check instead of directly
returning since there is a chance the function may tail call instead of
doing a return.

Also refactored the register usage audit code in to a separate function
to prepare for reuse in parsing linked shared libraries as well.