Implement the strategy of saving the conflicting register state in instrumented functions.
This backtracks from lazy and fine granular call graph analysis based
register usage and instrumentation since it is not precise enough with
the existance of unresolvable indirect calls. Now we are eagerly
calculating register usage of each function and saving the conflicting
register context around functions. We assume a fixed size SIMD stack
with all of extended avx2 registers as being reserved for the stack.