Significant changes to the instrumentation:
1. Replace spill stack and context stack with CFI context and user
context for vector registers. See function SharedLibraryInstrumentation
in src/instrument.cc for details.

2. User initial-exec TLS access model to avoid calling tls_get_addr,
which may cause infinite recursion when tls_get_addr is also
instrumented. See https://www.uclibc.org/docs/tls.pdf for elf's
definitions of four different TLS models.

3. It is difficult to determine which function is the "REAL" start of
the program execution. So, instead of struggline finding that function
to initialize our overflow stack, I change the code to initialize the
overflow stack on demand: in litecfi_overflow_stack_push, if we find
overflow_stack is still NULL, we initialize it.

4. The current implementation of libtls.so will cause libstdc++.so to be
one of the dependency of libtls.so. My guess is that bazel uses g++ to
compile libtls.so. However, if the application is C
only, then we will not instrument libstdc++.so. This causes an
uninstrumented libstdc++.so to be executed and initialized when libtls.so
is initialized. I have changed thread_local.cc to thread_local.c and
made necessary fixes to make sure libtls.so can be generated by gcc. So,
I can manually use gcc to generate libtls.so without using bazel.
However, I cannot make bazel to compile libtls.so with gcc. libtls.so
will still have libstdc++.so as its dependency.

TODO: it is better to make sure bazel use gcc to compile libtls.so. I
always invoke bazel as "CC=gcc CXX=g++ ./bazel.sh build"

5. Right now I disabled the use of "setup_memory" function to
initialize overflow_stack because it will call getpagesize, mmap, and
mprotect in libc, which are instrumented. So, obviously, calling
instrumented function during overflow_staci initialization is dangerous.
So, currently I statically allow space in TLS for overflow_stack.

6. I augmented spill_test and spill_and_overflow_test