Right now we are good with small tests, but still need to test with shared library instrumentation.
1. Context save & restore instrumentation uses BPatch
interfaces. It turns out that the base tramp for BPatch
instrumentation is going to enclose the PatchAPI level
instrumentation. So, if we need to save context, PatchAPI
level stack call does not need to set up r10 for RA location.

2. Add libstack_xmm.S file. This file contains edits from
the generated .S file. It is getting hard to change our source
code to change .S, mainly due to asmjit does not allow directly
emit an assembly as a string.

For this version of libstack_xmm.S, both stack push slot
and stack pop slot are 64-byte aligned. We can then easily
move up & down by pointers by plus and minus 64.

I tried to minimize GPR usages. For non-overflow case, we
use r10 for RA location and r11 for a couple of things.

3. Change overflow and context save code accordingly.
In libstack_xmm.S, to call overflow functions, we need to
save and restore rax, rcx, rdx because they are modified by
overflow functions. I think it is beneficial to only save &
restore them in the overflow slot, so that fast path does not
incur the overhead.

For overflow functions, r10 seems to be clobbered by the loader
when we call overflow functions the first time. So, I have to
use rdi to pass RA locations.

For overflow push, if the overflow stack is empty, it needs
to move the pop pointer to the overflow slot. Similarly, for
overflow pop, if after the pop, the overflow stack is empty,
it needs to move the pop pointer back to the last normal slot