Update the semantics of --shadow_stack flag.
We do not have avx based shadow stack versions anymore. So specifying
'mem' is reduandent. However we still need two versions where do the
full instrumentation and partial instrumentation based on static
analysis updated the flag to reflect that and refactored the code paths
to have a full instrumentation version.