Add initial static analysis passes for instrumentation lowering.
Currently I've verified up to the lowering pass with some toy
applications which exercises various control flow graphs featuring
conditionals and loops. Instrumentation coalescing passes are yet to be
tested. Without instrumentation coalescing the lowering is suboptimal in
that it will generate lot of redundant stack push nodes. Also the unsafe
loop handling could be more optimal if the loop header check can be
duplicated. But it is not done at the moment.